use About Us version of Video & VideoSection across website
