+++ src/pages/case-study/canon-giving-stories-a-new-form.js
@@ -17,6 +17,7 @@ import landscape from '../../images/case-study/at_the_heart_of_a_story.svg'
 import Image from '../../components/Common/Image'
 import { makeText } from '../../utils/makeText'
 import Head from '../../components/Common/Head'
+import VideoSection from '../../components/Common/VideoSection'
 
 const MobileOnly = styled.div`
   ${props => breakpoint(props.tablet ? 'tablet' : 'smallTablet')`
@@ -89,19 +90,6 @@ const GradientContent = ({ text, image }) => (
       </CenteredCol>
     </Row>
   </Fragment>
-)
-
-const Video = () => (
-  <div className="video-container">
-    <iframe
-      width="844"
-      height="480"
-      src="https://www.youtube.com/embed/MPPk-BkImsc"
-      frameBorder="0"
-      allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-      allowFullScreen
-    />
-  </div>
 )
 
 const IndexPage = ({
@@ -193,18 +181,14 @@ const IndexPage = ({
       </Grid>
 
       <MobileOnly>
-        <Video />
+        <VideoSection videoLink="https://www.youtube.com/embed/MPPk-BkImsc" />
       </MobileOnly>
 
       <GreyBackgroundOffset topMargin topOffset={-150}>
         <NoMobile>
           <Padding top={{ smallTablet: 0, tablet: 3 }} bottom={5}>
             <Grid>
-              <Row>
-                <Col width={[1]}>
-                  <Video />
-                </Col>
-              </Row>
+              <VideoSection videoLink="https://www.youtube.com/embed/MPPk-BkImsc" />
             </Grid>
           </Padding>
         </NoMobile>
